Fixed Knowledge Pick Button
